Refctor codice homePage
